refactor: extract shared ICE options parsing helpers across JS, Android, and iOS layers
